requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.2.post1
click==8.1.7
fastapi==0.115.2
greenlet==3.1.1
gunicorn==21.2.0
h11==0.14.0
idna==3.10
Mako==1.3.5
MarkupSafe==3.0.1
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.40.0
typing_extensions==4.12.2
uvicorn==0.32.0